refactor: 계층 책임에 따라 dto 분리
