+++ src/main/java/develop/grassserver/auth/application/dto/TokenDTO.java
@@ -1,0 +1,10 @@
+package develop.grassserver.auth.application.dto;
+
+import lombok.Builder;
+
+@Builder
+public record TokenDTO(
+        String refreshToken,
+        String accessToken
+) {
+}
