+++ src/main/java/develop/grassserver/auth/presentation/dto/TokenResponse.java
@@ -3,7 +3,7 @@ package develop.grassserver.auth.presentation.dto;
 import lombok.Builder;
 
 @Builder
-public record TokenDTO(
+public record TokenResponse(
         String refreshToken,
         String accessToken,
         String email
